chore: refactor and clean up code
