+++ src/components/styled/Wrappers.tsx
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+export const WelcomeWrapper = styled.div`
+  border-bottom: 3px dashed #ebda39;
+  padding-bottom: 2rem;
+  margin-bottom: 2rem;
+`;
+
+export const FooterWrapper = styled.footer`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: auto;
+  padding: 1rem 0.5rem;
+  margin-top: 2rem;
+  border-top: 3px dashed #ebda39;
+`;
