+++ src/loaders/latestAdsLoader.ts
@@ -4,6 +4,7 @@ import { getLatestAds } from '../services/adService';
 export type LatestAdsLoader = {
   ads: Ad[];
 };
+
 export const latestAdsLoader = async () => {
   const data = await getLatestAds();
