+++ src/models/AdExt.ts
@@ -51,17 +51,10 @@ export interface AdExt extends Ad {
     }[];
   };
 
-  // workplace_address: {
-  //   region?: string;
-  //   municipality?: string;
-  // };
-
   application_contacts?: {
     name?: string;
     description?: string;
     email?: string;
     telephone?: string;
   }[];
-
-  // publication_date?: string;
 }
