+++ src/services/adService.ts
@@ -9,21 +9,7 @@ const SEARCH = 'search?occupation-group=DJh5_yyF_hEM';
 // occupation-group=DJh5_yyF_hEM targets jobs within the software development career.
 // offset means where to start in the list and limit is how many to fetch.
 
-export const getAds = async (query: string) => {
-  const encodedQuery = encodeURIComponent(query);
-  const data = await get<APIResponse>(
-    `${BASE_URL}${SEARCH}&q=${encodedQuery}&offset=0&limit=100`
-  ); // offset part will be dynamic so will have to be replaced
-
-  return data.hits;
-};
-
-// Under construction
-export const getAdsNew = async (
-  query: string,
-  offset: number,
-  limit: number
-) => {
+export const getAds = async (query: string, offset: number, limit: number) => {
   const encodedQuery = encodeURIComponent(query);
   const response = await get<APIResponse>(
     `${BASE_URL}${SEARCH}&q=${encodedQuery}&offset=${offset}&limit=${limit}`
